+++ src/acvp_kas_ffc.c
@@ -311,7 +311,7 @@ static ACVP_RESULT acvp_kas_ffc_comp(ACVP_CTX *ctx,
             ACVP_LOG_INFO("Found new KAS-FFC Component test vector...");
             testval = json_array_get_value(tests, j);
             testobj = json_value_get_object(testval);
-            tc_id = (unsigned int)json_object_get_number(testobj, "tcId");
+            tc_id = json_object_get_number(testobj, "tcId");
 
             eps = json_object_get_string(testobj, "ephemeralPublicServer");
             if (!eps) {
